+++ Sources/Transifex/Core.swift
@@ -164,19 +164,38 @@ class NativeCore : TranslationProvider {
     /// - Parameter status: An optional status so that only strings matching translation status are
     /// fetched.
     /// - Parameter completionHandler: The completion handler that informs the caller with the
-    /// new translations and a list of possible errors that might have occured
+    /// new translations and a list of possible errors that might have occured. The completion handler is
+    /// called from a background thread.
     func fetchTranslations(_ localeCode: String? = nil,
                            tags: [String] = [],
                            status: String? = nil,
                            completionHandler: TXPullCompletionHandler? = nil) {
         cdsHandler.fetchTranslations(localeCode: localeCode,
                                      tags: tags,
-                                     status: status) { (translations, errors) in
+                                     status: status) { [weak self] (translations, errors) in
+            guard let self = self else {
+                return
+            }
+
             if errors.count > 0 {
                 Logger.error("\(#function) Errors: \(errors)")
             }
 
-            self.cache.update(translations: translations)
+            // Only update the cache if the translations structure is not
+            // empty, avoiding cases where no translations were fetched due
+            // to errors (e.g. network offline etc).
+            // We do not check for errors here as some translations might
+            // have failed and some others might not. As long as the
+            // translations structure is not empty, we are OK to update the
+            // cache (and, by extension, the stored file).
+            if !translations.isEmpty {
+                // Update cache using the fetched translations in main thread
+                // to ensure proper cache updates for custom implemented caching
+                // solutions.
+                DispatchQueue.main.async {
+                    self.cache.update(translations: translations)
+                }
+            }
 
             completionHandler?(translations, errors)
         }
@@ -191,22 +210,33 @@ class NativeCore : TranslationProvider {
     ///   - completionHandler: A callback to be called when the push operation is complete with a
     /// boolean argument that informs the caller that the operation was successful (true) or not (false) and
     /// an array that may or may not contain any errors produced during the push operation and an array of
-    /// non-blocking errors (warnings) that may have been generated during the push procedure.
+    /// non-blocking errors (warnings) that may have been generated during the push procedure. The
+    /// completion handler is called from a background thread.
     func pushTranslations(_ translations: [TXSourceString],
                           configuration: TXPushConfiguration = TXPushConfiguration(),
                           completionHandler: @escaping (Bool, [Error], [Error]) -> Void) {
         cdsHandler.pushTranslations(translations,
-                                    configuration: configuration,
-                                    completionHandler: completionHandler)
+                                    configuration: configuration) { [weak self]
+            success, errors, warnings in
+            guard let _ = self else {
+                return
+            }
+            completionHandler(success, errors, warnings)
+        }
     }
     
     /// Forces CDS cache invalidation.
     ///
     /// - Parameter completionHandler: A callback to be called when force cache invalidation is
     /// complete with a boolean argument that informs the caller that the operation was successful (true) or
-    /// not (false).
+    /// not (false).  The completion handler is called from a background thread.
     func forceCacheInvalidation(completionHandler: @escaping (Bool) -> Void) {
-        cdsHandler.forceCacheInvalidation(completionHandler: completionHandler)
+        cdsHandler.forceCacheInvalidation { [weak self] success in
+            guard let _ = self else {
+                return
+            }
+            completionHandler(success)
+        }
     }
     
     /// Used by the Swift localizedString(format:arguments:) methods found in the
@@ -361,7 +391,7 @@ render '\(stringToRender)' locale code: \(localeCode) params: \(params). Error:
 /// A static class that is the main point of entry for all the functionality of Transifex Native throughout the SDK.
 public final class TXNative : NSObject {
     /// The SDK version
-    internal static let version = "2.0.4"
+    internal static let version = "2.0.5"
     
     /// The filename of the file that holds the translated strings and it's bundled inside the app.
     public static let STRINGS_FILENAME = "txstrings.json"
@@ -566,6 +596,7 @@ token: \(token)
     ///   - status: An optional status so that only strings matching translation status are fetched.
     ///   - completionHandler: The completion handler that informs the caller when the operation
     ///   is complete, reporting the new translations and a list of possible errors that might have occured.
+    ///   The completion handler is called from a background thread.
     @objc
     public static func fetchTranslations(_ localeCode: String? = nil,
                                          tags: [String]? = nil,
@@ -586,7 +617,8 @@ token: \(token)
     ///   - completionHandler: A callback to be called when the push operation is complete with a
     /// boolean argument that informs the caller that the operation was successful (true) or not (false) and
     /// an array that may or may not contain any errors produced during the push operation and an array of
-    /// non-blocking errors (warnings) that may have been generated during the push procedure.
+    /// non-blocking errors (warnings) that may have been generated during the push procedure. The
+    /// completion handler is called from a background thread.
     @objc
     public static func pushTranslations(_ translations: [TXSourceString],
                                         configuration: TXPushConfiguration = TXPushConfiguration(),
@@ -600,7 +632,7 @@ token: \(token)
     ///
     /// - Parameter completionHandler: A callback to be called when force cache invalidation is
     /// complete with a boolean argument that informs the caller that the operation was successful (true) or
-    /// not (false).
+    /// not (false).  The completion handler is called from a background thread.
     @objc
     public static func forceCacheInvalidation(completionHandler: @escaping (Bool) -> Void) {
         tx?.forceCacheInvalidation(completionHandler: completionHandler)